requirements.txt
customtkinter==5.2.2
darkdetect==0.8.0
greenlet==3.4.0
packaging==26.1
pillow==12.2.0
SQLAlchemy==2.0.49
typing_extensions==4.15.0
